% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmbn_body.R
\docType{data}
\name{gmbn_body}
\alias{gmbn_body}
\title{Gaussian mixture Bayesian network learned from the NHANES body composition
dataset}
\format{
A \code{gmbn} object.
}
\usage{
gmbn_body
}
\description{
This Gaussian mixture dynamic Bayesian network is learned from the NHANES
body composition dataset, following the example provided in the documentation
page of function \code{\link{struct_learn}}.
}
\seealso{
\code{\link{data_air}}, \code{\link{data_body}},
\code{\link{gmdbn_air}}, \code{\link{gmm_body}}
}
\keyword{datasets}
