% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmdbn_air.R
\docType{data}
\name{gmdbn_air}
\alias{gmdbn_air}
\title{Gaussian mixture dynamic Bayesian network learned from the Beijing air
quality dataset}
\format{
A \code{gmdbn} object.
}
\usage{
gmdbn_air
}
\description{
This Gaussian mixture dynamic Bayesian network is learned from the Beijing
air quality dataset, following the example provided in the documentation page
of function \code{\link{struct_learn}}.
}
\seealso{
\code{\link{data_air}}, \code{\link{data_body}},
\code{\link{gmbn_body}}, \code{\link{gmm_body}}
}
\keyword{datasets}
