% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_body.R
\docType{data}
\name{gmm_body}
\alias{gmm_body}
\title{Gaussian mixture model learned from the NHANES body composition dataset}
\format{
A \code{gmm} object.
}
\usage{
gmm_body
}
\description{
This Gaussian mixture model is learned from the NHANES body composition
dataset, following the example provided in the documentation page of function
\code{\link{stepwise}}.
}
\seealso{
\code{\link{data_air}}, \code{\link{data_body}},
\code{\link{gmbn_body}}, \code{\link{gmdbn_air}}
}
\keyword{datasets}
