% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network}
\alias{network}
\title{Display the graphical structure of a Gaussian mixture Bayesian network}
\usage{
network(gmbn)
}
\arguments{
\item{gmbn}{An object of class \code{gmbn}.}
}
\value{
A \code{visNetwork} object displaying the graphical structure.
}
\description{
This function displays the graphical structure of a Gaussian mixture
Bayesian network.
}
\examples{
data(gmbn_body)
network(gmbn_body)

data(gmdbn_air)
network(gmdbn_air$b_2)

}
