% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_var.R
\name{remove_var}
\alias{remove_var}
\title{Remove variables from a Gaussian mixture model}
\usage{
remove_var(gmm, var)
}
\arguments{
\item{gmm}{An object of class \code{gmm}.}

\item{var}{A character vector containing the removed variables.}
}
\value{
The \code{gmm} object after removing the variables.
}
\description{
This function removes variables from a Gaussian mixture model.
}
\examples{
data(gmm_body)
gmm_1 <- remove_var(gmm_body, "FAT")

}
\seealso{
\code{\link{add_var}}, \code{\link{rename_var}}
}
