% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_var.R
\name{rename_var}
\alias{rename_var}
\title{Rename variables of a Gaussian mixture model}
\usage{
rename_var(gmm, var, names)
}
\arguments{
\item{gmm}{An object of class \code{gmm}.}

\item{var}{A character vector containing the renamed variables.}

\item{names}{A character vector containing the respective new names of the
variables.}
}
\value{
The \code{gmm} object after renaming the variables.
}
\description{
This function renames variables of a Gaussian mixture model.
}
\examples{
data(gmm_body)
gmm_1 <- rename_var(gmm_body, "FAT", "BODY_FAT")

}
\seealso{
\code{\link{add_var}}, \code{\link{remove_var}}
}
